Fix explain analyze with CTE inserts (#1)
* Fix crash on explain analyze with insert CTE

Discovered a crash when doing an explain analyze on a CTE containing an insert that did not reference the CTE in the select statement. (ie `EXPLAIN ANALYZE WITH a as (INSERT ... ) SELECT 1`. 

This fixes by copying the eref from the parent hypertable to the chunk so that columns can be described. It will only copy the eref from the hypertable if there is a range table entry available.